requirements.txt
tkinter
langdetector